job-role introduced
After refactoring loading of json, there is now a first
directive job-role, to display the title of a job-box.
It can be feed by arbitrary strings. To prepare the code
for the usage in ng-repeat, there is a job-variable in scope
referring the first job loaded and will be deleted afterwards.